♻️ Address review feedback on plugin init and Shopify plugin tests
