+++ packages/browser-rum-shopify/src/domain/shopifyAnalytics.spec.ts
@@ -1,28 +1,6 @@
 import { TIMEOUT_ERROR_MESSAGE } from '@datadog/browser-core'
-import type { ShopifyAnalyticsApi, ShopifyPixelEvent } from './shopifyAnalytics'
+import { createFakeAnalytics, pageViewedEvent } from '../test/mockShopifyAnalytics'
 import { waitForPageViewedEvent } from './shopifyAnalytics'
-
-function createFakeAnalytics() {
-  const subscribers = new Map<string, (event: ShopifyPixelEvent) => void>()
-  const analytics: ShopifyAnalyticsApi = {
-    subscribe: jasmine.createSpy('subscribe').and.callFake((eventName: string, callback) => {
-      subscribers.set(eventName, callback)
-    }),
-  }
-  return {
-    analytics,
-    emit: (eventName: string, event: ShopifyPixelEvent) => subscribers.get(eventName)?.(event),
-  }
-}
-
-function pageViewedEvent(url: string | undefined): ShopifyPixelEvent {
-  return {
-    name: 'page_viewed',
-    id: '1',
-    timestamp: '2026-07-06T00:00:00Z',
-    context: { document: { location: { href: url } } },
-  }
-}
 
 describe('waitForPageViewedEvent', () => {
   it('resolves with the page_viewed event once it is emitted', async () => {
